requirements.txt
fastapi
uvicorn
pydantic
gTTS
Pillow
moviepy
python-dotenv
Jinja2